--- resources/js/pages/dashboard.tsx
@@ -1,30 +1,0 @@
-import { AppSidebar } from "@/Components/app-sidebar"
-import { ChartAreaInteractive } from "@/Components/chart-area-interactive"
-import { DataTable } from "@/Components/data-table"
-import { SectionCards } from "@/Components/section-cards"
-import { SiteHeader } from "@/Components/site-header"
-import { SidebarInset, SidebarProvider } from "@/Components/ui/sidebar"
-
-import data from "./data.json"
-
-export default function Page() {
-  return (
-    <SidebarProvider>
-      <AppSidebar variant="inset" />
-      <SidebarInset>
-        <SiteHeader />
-        <div className="flex flex-1 flex-col">
-          <div className="@container/main flex flex-1 flex-col gap-2">
-            <div className="flex flex-col gap-4 py-4 md:gap-6 md:py-6">
-              <SectionCards />
-              <div className="px-4 lg:px-6">
-                <ChartAreaInteractive />
-              </div>
-              <DataTable data={data} />
-            </div>
-          </div>
-        </div>
-      </SidebarInset>
-    </SidebarProvider>
-  )
-}
